vibra-periodic: Use same range as in rumble and periodic
We use [0.0, 1.0] in the other haptic feedbacks so let's make periodic
do the same.

We adjust dev-vibra to match that.

Closes: https://source.puri.sm/Librem5/feedbackd/-/issues/86

+++ src/fbd-dev-vibra.h
@@ -19,8 +19,11 @@ gboolean     fbd_dev_vibra_rumble (FbdDevVibra *device,
                                    double       magnitude,
                                    guint        duration,
                                    gboolean     upload);
-gboolean     fbd_dev_vibra_periodic (FbdDevVibra *self, guint duration, guint magnitude,
-				     guint fade_in_level, guint fade_in_time);
+gboolean     fbd_dev_vibra_periodic (FbdDevVibra *self,
+                                     guint        duration,
+                                     double       magnitude,
+                                     double       fade_in_level,
+                                     guint        fade_in_time);
 gboolean     fbd_dev_vibra_stop (FbdDevVibra *self);
 gboolean     fbd_dev_vibra_remove_effect (FbdDevVibra *self);
 GUdevDevice *fbd_dev_vibra_get_device(FbdDevVibra *self);
